requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numpy==2.2.6
pillow==11.2.1
propcache==0.3.1
python-dotenv==1.1.0
pytz==2025.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.0
yarl==1.20.0
storage3==0.5.2
supabase==1.0.3
requests==2.32.3
faiss-cpu==1.9.0
